implement chronicle rehearsal
