+++ src/main/resources/data/test_chronicle_dict_data.csv
@@ -1,0 +1,7 @@
+id,fullname, title
+1,com.android.dialer,Android Phone Dialer?
+2,com.android.chrome,Google Chrome: Fast & Secure
+3,com.google.android.apps.maps,Google Maps
+4,com.google.android.gm,Gmail
+5,com.android.camera2,Camera
+6,com.google.android.youtube,YouTube
