+++ src/main/resources/data/test_chronicle_empty_data.csv
@@ -1,0 +1,1 @@
+,app_to_device_id,app_to_participants_id,application_data_dt,application_data_fullname,application_data_id,application_data_recordtype,application_data_title,application_data_tz,device_id,device_model,device_to_study_id,device_to_user_id,device_version,study_description,study_email,study_fullname,study_id,study_name,study_to_participants_id,study_version,user,user_0,user_1,user_2,user_apps_fullname,user_apps_id,user_apps_to_device_id,user_apps_to_user_id,user_apps_user
