+++ src/main/resources/flights/test_chronicle_dict_flight.yaml
@@ -1,0 +1,16 @@
+organizationId: 00000000-0000-0001-0000-000000000000
+associationDefinitions: {}
+entityDefinitions:
+  application_dictionary:
+    entitySetName: chronicle_application_dictionary
+    fqn: ol.applicationdata
+    propertyDefinitions:
+      general.fullname:
+        type: general.fullname
+        column: fullname
+      ol.title:
+        type: ol.title
+        column: title
+      general.stringid:
+        type: general.stringid
+        column: id
